### 2023-11-21 #6078 @editorialbot
[PRE REVIEW]: pylattica: a package for prototyping lattice models in chemistry and materials science
**Submitting author:** @mcgalcode (<a href="http://orcid.org/0009-0008-4099-6144">Max C. Gallant</a>)
**Repository:** https://github.com/mcgalcode/pylattica
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.2
