### 2023-11-21 #6078 @Kevin-Mattheus-Moerman
@mcgalcode Thanks for this submission. I am the AEiC for this track, and here to help process initial steps. I have started to look for a handling editor. Note that many are handling other submissions so it may take some time before we are able to assign one.

For the moment, could you please try to address the above potentially missing/invalid DOIs? :point_up:

You can call `@editorialbot generate pdf` to update the paper, and you can call `@editorialbot check references` to check those DOIs again. Thanks.
