+++ paper/paper.bib
@@ -17,18 +17,17 @@
 	file = {1-s2.0-S0927025612005435-main.pdf:/Users/mgallant/Documents/school/articles/1-s2.0-S0927025612005435-main.pdf:application/pdf;ScienceDirect Snapshot:/Users/mgallant/Zotero/storage/YIGIBQA4/S0927025612005435.html:text/html},
 }
 
-@misc{morgan_lattice_mc_2023,
-	title = {lattice\_mc: {A} {Python} {Lattice}-{Gas} {Monte} {Carlo} {Module}},
-	copyright = {MIT},
-	shorttitle = {lattice\_mc},
-	url = {https://github.com/bjmorgan/lattice_mc},
-	abstract = {Lattice gas Monte Carlo simulation code},
-	urldate = {2023-10-23},
-	author = {Morgan, Benjamin},
-	month = aug,
-	year = {2023},
-	note = {original-date: 2016-12-06T16:40:48Z},
-	keywords = {kinetic-monte-carlo, lattice-gas, monte-carlo, monte-carlo-simulation, solid-state-physics},
+@article{Morgan2017,
+	doi = {10.21105/joss.00247},
+	url = {https://doi.org/10.21105/joss.00247},
+	year = {2017},
+	publisher = {The Open Journal},
+	volume = {2},
+	number = {13},
+	pages = {247},
+	author = {Benjamin J. Morgan},
+	title = {lattice_mc: A Python Lattice-Gas Monte Carlo Module},
+	journal = {Journal of Open Source Software}
 }
 
 @misc{feistenauer_cellular_automaton_2021,
@@ -206,6 +205,7 @@ Program title: KMCLib Catalogue identifier: AESZ\_v1\_0 Program summary URL:http
 	title = {A {Practical} {Guide} to {Surface} {Kinetic} {Monte} {Carlo} {Simulations}},
 	volume = {7},
 	issn = {2296-2646},
+	doi = {10.3389/fchem.2019.00202},
 	url = {https://www.frontiersin.org/articles/10.3389/fchem.2019.00202},
 	abstract = {This review article is intended as a practical guide for newcomers to the field of kinetic Monte Carlo (KMC) simulations, and specifically to lattice KMC simulations as prevalently used for surface and interface applications. We will provide worked out examples using the kmos code, where we highlight the central approximations made in implementing a KMC model as well as possible pitfalls. This includes the mapping of the problem onto a lattice and the derivation of rate constant expressions for various elementary processes. Example KMC models will be presented within the application areas surface diffusion, crystal growth and heterogeneous catalysis, covering both transient and steady-state kinetics as well as the preparation of various initial states of the system. We highlight the sensitivity of KMC models to the elementary processes included, as well as to possible errors in the rate constants. For catalysis models in particular, a recurrent challenge is the occurrence of processes at very different timescales, e.g., fast diffusion processes and slow chemical reactions. We demonstrate how to overcome this timescale disparity problem using recently developed acceleration algorithms. Finally, we will discuss how to account for lateral interactions between the species adsorbed to the lattice, which can play an important role in all application areas covered here.},
 	urldate = {2023-10-24},
