### 2023-11-21 #6078 @mcgalcode
Hi @Kevin-Mattheus-Moerman, I appreciate your help with the review of my paper! This all sounds good to me.

I fixed the first of the two DOIs, but the third suggestion refers to something unrelated. That bibliography entry is for an article in Scientific American on jstor, which doesn't seem to have a DOI associated with it (https://www.jstor.org/stable/24927642).
