### 2023-11-21 #6078 @Kevin-Mattheus-Moerman
@mcgalcode these "calls" (e.g. ` @editorialbot check references`) to the bot need to be the first thing (and perhaps on their own) in a comment
