### 2023-12-20 91e2857
Fix growth API and improve some serialization
