+++ tests/structures/square_grid/test_growth_setup.py
@@ -10,7 +10,11 @@ def test_growth_setup():
     simulation_side_length = 20
     total_num_sites = 4
     background_phase = "A"
-    site_phases = ["B", "C"]
+
+    nuc_amts = {
+        'B': 1,
+        'C': 1
+    }
     buffer = 2 # Each site should be at least 2 cells away from any other
 
     growth_setup = GrowthSetup(phases)
@@ -18,7 +22,7 @@ def test_growth_setup():
         simulation_side_length,
         background_spec=background_phase,
         num_sites_desired=total_num_sites,
-        nuc_species=site_phases,
+        nuc_amts=nuc_amts,
         buffer=buffer,
         nb_builder=MooreNbHoodBuilder(1)
     )
