### 2023-12-21 #6170 @editorialbot
[REVIEW]: pylattica: a package for prototyping lattice models in 2 chemistry and materials science
**Submitting author:** @mcgalcode (<a href="http://orcid.org/0009-0008-4099-6144">Max C. Gallant</a>)
**Repository:** https://github.com/mcgalcode/pylattica
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.3
