### 2024-01-11 mcgalcode/pylattica#12 @RiesBen
Review Repository
Hi ,
I just added a few comments here, regarding my review of your JOSS submission.

The package implements lattice-based simulation approaches for prototyping methods. The code is very modular and dynamic, allowing a wide range of uses for custom simulation types and systems. The documentation very well presents different possible use cases.

Repository:
- The License you provide is not OSI-approved, right? But this is a requirement for JOSS publications.
- Is there a general use for having the  .vscode code in the repository or is this an artifact?
- Many functions and classes are well documented. However some docstrings are missing in neighborhood_builder,  structure_builder, in the visualization module, and the square_grid setup. Adding those would make the usage of functions and classes clearer for new users of the package

Documentation:
- It would be great to have a general install section in your readme and documentation for users to quickly get the package installed. You can add the PyPi install variant and maybe also how to run the unit tests locally, too.
- I found a jupyter notebook in the docs, that uses outdated code, an update on the code would be good. (https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/docs/examples/life_like_demos.ipynb)

This is more a detailed thing:
 - while executing one example, I observed, that the labels of the simulation visualization started to change their order (1.dead, 2. alive - 1. dead, 2. alive). Maybe it is possible to fix their order for all frames?
