### 2024-01-11 #6170 @RiesBen
Hi,
I added code and Documentation Review here:
[Review Repository](https://github.com/mcgalcode/pylattica/issues/12)
