### 2024-01-25 mcgalcode/pylattica#13 @amkrajewski
Code Review
Here is a list of comments regarding the software for my review of your JOSS submission. I will open a separate issue for the manuscript to keep things organized.

- **The review was a pleasure to complete. _pylattica_ is well written, and its codebase is clearly organized into a directory structure**, which is relatively easy to navigate thanks to the good naming of things.

- Installation instructions are missing and should be added for novice Python users, but a simple `pip install -e .` everything went smoothly for me on a clean VM I've set up to play with it. All tests passed in just under 20s.

- The documentation contains several neat, runnable examples (guides) that appear to be Jupyter notebooks. Placing them in a dedicated directory in the repository would be great. I couldn't find them currently.

- It's great to see that functions (almost) always have type hints correctly following PEP 484, significantly improving end-user experience and reducing the possibility of incorrect use, especially considering numerous custom classes.

- The coverage is tested with Codecov, and it's 98, which is excellent. A minor issue is that the badge in the README does not link to the actual report.

- Tests are run on the Ubuntu Linux with Python 3.8-10. Authors may want to bump that to 3.9-11, as Python 3.8 is about to reach end-of-life in a couple of months, while Python 3.11 is the oldest version actively receiving non-security bug fixes. Adding Windows and Mac to the test matrix would also be nice.

- Docstrings are occasionally missing in important places: https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/structures/square_grid/growth_setup.py#L7 or have been created as templates with no information
https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/structures/square_grid/grid_setup.py#L177-L181

- In some regions of the codebase, Python classes seem to be defined just to provide different inputs such as simple string values https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/models/game_of_life/controller.py#L51-L52
to the procedure running the simulation. I can see the usefulness in more complex cases, but it may be undesirable here as it obscures usage from the end customers and crowds the package namespace.

- Relative https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/models/growth/growth_controller.py#L1 and absolute https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/models/game_of_life/controller.py#L1 imports are not used consistently decreasing readability.

- [Google](https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/structures/square_grid/grid_setup.py#L26) and [NumPy](https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/discrete/discrete_state_result_analyzer.py#L30) docstring styles are mixed / not used consistently. It may be the reason [parts of the documentation](https://mcgalcode.github.io/pylattica/reference/core/lattice/) are not processed correctly.

- In [parts of the documentation](https://mcgalcode.github.io/pylattica/guides/neighborhoods/), some images do not display correctly (for me), and some links, e.g., to the CHANGELOG, are broken.
