### 2024-01-25 mcgalcode/pylattica#14 @amkrajewski
Manuscript Review
Here is a list of comments regarding the manuscript (living in the [paper branch](https://github.com/mcgalcode/pylattica/tree/paper)) for my review of your JOSS submission.

- **The manuscript reads well, both in terms of contents and style. It introduces the core functionality of _pylattica_ software and explains the motivation for creating an extensible platform that can help prototype new methods and, at the same time, keep them consistent with each other to improve reusability.**

- The interaction with `pymatgen` and the degree to which it underlies the software seems a bit glossed over in the manuscript. In particular, [parts dealing with lattice](https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/core/lattice.py) are largely directly copied from there, which is perfectly ok and great for extensibility, but (1) should be explicitly mentioned in publication and (2) `pymatgen`'s MIT license notice has to be attached to the codebase.

- *(L17)* - Typo? It seems it should have been "state of the neighborhood."

- *(L42-46)* - These two sentences are a bit confusing and also slightly redundant in the information they carry. A quick rewriting from scratch may help.

- *(Sec. Simulation Execution*) - A simple flowchart would be helpful in visualizing how different classes interoperate.

- The way the neighborhood between sites is established should be discussed in more detail. The [current codebase](https://github.com/mcgalcode/pylattica/blob/9b6d8f43b0f269576b720826257ae2eb5652153b/src/pylattica/core/neighborhood_builders.py) is hard-coded to use either Motiffs or Cartesian distance sphere as the cutoff, and it may be a limiting factor in some cases. Allowing more custom Voronoi tesselation and other methods wrapped into the `pymatgen` structure analysis toolset may be more future-proof.

- *(L93)* - The "SimulationStep" is (1) not a monospace `code` block, breaking style from the rest of the manuscript, and has not been introduced earlier, nor did I see it in the code. Perhaps an error? `SimulationState` seemingly fits here.
