### 2024-01-25 #6170 @amkrajewski
Hi @richardjgowers and @RiesBen, I finished my review. I opened two separate issues: one for code at https://github.com/mcgalcode/pylattica/issues/13 and one for the manuscript at https://github.com/mcgalcode/pylattica/issues/14, just to keep things organized.
