### 2024-02-07 mcgalcode/pylattica#14 @mcgalcode
Hi @amkrajewski! Thanks for your review - I'm putting your comments in checklist form here so I can keep track of my progress in addressing them.

- [ ] Mention c&p lattice distance methods in paper
- [ ] Move to MIT license
- [ ] Fix L17 typo
- [ ] Revise L42-46
- [ ] Flowchart visualization (agreed this would be helpful!)
- [ ] Expand description of neighborhood construction
- [ ] Fix L93 monospace + typo issue
