### 2024-02-07 mcgalcode/pylattica#12 @mcgalcode
Hi @RiesBen - thanks for your review! I'm just adding a checklist here so I can keep track of my progress addressing your comments.

- [ ] Fix License (have to switch to pymatgen MIT license anyway)
- [ ] Remove .vscode artifact
- [ ] Fill in missing documentation
- [ ] Add installation instructions
- [ ] Fix the life_like_demos notebook
- [ ] Fix ordering in 2D viz labels
