### 2024-02-07 mcgalcode/pylattica#13 @mcgalcode
Hello again @amkrajewski! Another checklist for my own reference:

- [ ] Add instructions for running tests
- [ ] Either move or clarify location of example notebooks
- [ ] Readme codecov badge link fix
- [ ] At the very least, bump python version in testing matrix, try adding windows + mac as well
- [ ] Add doc strings in places where they are missing
- [ ] Identify classes that could be enums and refactor
- [ ] Unify imports to relative format
- [ ] Unify docstring styles
- [ ] Fix broken images and links in documentation
