### 2024-02-07 f262fbf
Add installation and testing instructions for development
