### 2024-02-07 e6c8351
Add installation instructions
