### 2024-02-07 ed4cae1
Specify location of jupyter notebooks in README
