### 2024-02-07 a1b567e
Fix link and add link to guide source notebooks
