### 2024-02-07 b2bd19d
Only use parallel runner where required, and skip if windows is platform
