+++ tests/visualization/test_square_grid_artists.py
@@ -14,7 +14,7 @@ def test_step_artist():
     setup = DiscreteGridSetup(phases)
     simulation = setup.setup_noise(10, ["dead", "alive"])
     controller = Life(structure = simulation.structure)
-    runner = SynchronousRunner(parallel=True)
+    runner = SynchronousRunner(parallel=False)
     result = runner.run(simulation.state, controller, 10, verbose=False)
     cell_artist = DiscreteCellArtist.from_discrete_state(result.last_step)
     artist = SquareGridArtist2D(simulation.structure, cell_artist)
@@ -27,7 +27,7 @@ def test_result_artist():
     setup = DiscreteGridSetup(phases)
     simulation = setup.setup_noise(10, ["dead", "alive"])
     controller = Life(structure = simulation.structure)
-    runner = SynchronousRunner(parallel=True)
+    runner = SynchronousRunner(parallel=False)
     result = runner.run(simulation.state, controller, 10, verbose=False)
     cell_artist = DiscreteCellArtist.from_discrete_result(result)
     step_artist = SquareGridArtist2D(simulation.structure, cell_artist)
@@ -50,7 +50,7 @@ def test_step_artist_3D():
     phases = PhaseSet(["dead", "alive"])
     setup = DiscreteGridSetup(phases, dim=3)
     simulation = setup.setup_noise(3, ["dead", "alive"])
-    runner = SynchronousRunner(parallel=True)
+    runner = SynchronousRunner(parallel=False)
     result = runner.run(simulation.state, SimpleController(), 4, verbose=False)
 
     cell_artist = DiscreteCellArtist.from_discrete_state(result.last_step)
