+++ pyproject.toml
@@ -105,6 +105,8 @@ known_first_party = "pylattica"
 
 [tool.pytest]
 addopts = "--durations=10"
+norecursedirs = "tests/test_helpers"
+
 
 [tool.setuptools.package-data]
 pylattica = ["py.typed"]
