### 2024-02-07 c4478e8
Expand testing platforms and update python versions
