+++ .github/workflows/testing.yaml
@@ -40,8 +40,8 @@ jobs:
     strategy:
       max-parallel: 6
       matrix:
-        os: [ubuntu-latest]
-        python_version: ["3.8", "3.9", "3.10"]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python_version: ["3.9", "3.10", "3.11"]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v3
