### 2024-02-07 0dc6de4
Switch to numpy docstrings
