+++ src/pylattica/structures/square_grid/grid_setup.py
@@ -26,11 +26,15 @@ class DiscreteGridSetup:
     """
 
     def __init__(self, phase_set: PhaseSet, dim: int = 2):
-        """Initializes a laboratory object by providing a reaction set that is used to
+        """Initializes a DiscreteGridSetup object by providing a set of phases that is used to
         specify cell states
 
-        Args:
-            phase_set (PhaseSet):
+        Parameters
+        ----------
+        phase_set : PhaseSet
+            A PhaseSet instance containing the possible phases that might be assigned to cells.
+        dim : int, optional
+            The dimension of the simulations to setup, by default 2
         """
         if dim == 2:
             self._builder = SimpleSquare2DStructureBuilder()
@@ -56,13 +60,19 @@ class DiscreteGridSetup:
         """Generates a starting state that is divided into two phases. One phase
         occupies the left half of the state, and one phase occupies the right half of the state
 
-        Args:
-            size (int): The side length of the state
-            p1 (str): The name of the left phase
-            p2 (str): The name of the right phase
+        Parameters
+        ----------
+        size : int
+            The side length of the state
+        p1 : str
+            The name of the left phase
+        p2 : str
+            The name of the right phase
 
-        Returns:
-            SimulationState:
+        Returns
+        -------
+        Simulation
+            The resulting Simulation.
         """
         structure = self._builder.build(size=size)
         state: SimulationState = self._build_blank_state(structure)
@@ -80,14 +90,21 @@ class DiscreteGridSetup:
         """Generates a starting state with a bulk phase surrounding a particle in the
         center of the state.
 
-        Args:
-            size (int): The side length of the state
-            radius (int): The radius of the particle
-            bulk_phase (str): The name of the bulk
-            particle_phase (str): The name of the particle phase
+        Parameters
+        ----------
+        size : int
+            The side length of the state
+        radius : int
+            The radius of the particle
+        bulk_phase : str
+            The name of the bulk
+        particle_phase : str
+            The name of the particle phase
 
-        Returns:
-            SimulationState:
+        Returns
+        -------
+        Simulation
+            The resulting Simulation
         """
         structure = self._builder.build(size=size)
         state: SimulationState = self.setup_solid_phase(structure, bulk_phase)
@@ -110,15 +127,23 @@ class DiscreteGridSetup:
         """Generates a starting state with a one phase in the background and num_particles particles distributed
         onto it randomly
 
-        Args:
-            size (int): The size of the state
-            radius (int): The radius of the particles to drop
-            num_particles (int): The number of particles
-            bulk_phase (str): The name of the containing phase
-            particle_phases (str): The name of the particulate phase
+        Parameters
+        ----------
+        size : int
+            The size of the simulation
+        radius : int
+            The radius of the particles to drop
+        num_particles : int
+            The number of particles
+        bulk_phase : str
+            The name of the containing phase
+        particle_phases : str
+            The name of the particulate phase
 
-        Returns:
-            SimulationState:
+        Returns
+        -------
+        Simulation
+            The resulting Simulation.
         """
         structure = self._builder.build(size)
         state: SimulationState = self.setup_solid_phase(structure, bulk_phase)
@@ -149,7 +174,25 @@ class DiscreteGridSetup:
                 )
         return state
 
-    def setup_coords(self, size: int, background_state: str, coordinates: dict):
+    def setup_coords(self, size: int, background_state: str, coordinates: dict) -> Simulation:
+        """Generates a simulation filled with a specifed background phase, and specific
+        sites filled with specific phases as defined by the coordinates parameter.
+
+        Parameters
+        ----------
+        size : int
+            The size of the simulation to generate
+        background_state : str
[…]
