+++ src/pylattica/core/lattice.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import List, Tuple
+from typing import List, Tuple, Union
 
 import numpy as np
 from numpy.typing import ArrayLike
@@ -8,7 +8,22 @@ from .constants import OFFSET_PRECISION
 import math
 
 
-def periodize(frac_coords, periodic=True):
+def periodize(frac_coords, periodic: Union[Tuple[bool], bool] = True):
+    """Moves fractional coordinates into the unit cell.
+
+    Parameters
+    ----------
+    frac_coords : ArrayLike
+        The fractional coordinates to periodize.
+    periodic : bool, optional
+        Either a single boolean or a tuple of booleans indicating the periodicity
+        of each dimension, by default True
+
+    Returns
+    -------
+    ArrayLike
+        The periodized coordinates
+    """
     if not isinstance(periodic, tuple):
         periodic = [periodic for _ in frac_coords]
 
@@ -19,15 +34,21 @@ def pbc_diff_frac_vec(fcoords1: ArrayLike, fcoords2: ArrayLike, periodic):
     """Returns the 'fractional distance' between two coordinates taking into
     account periodic boundary conditions. (from pymatgen)
 
-    Args:
-        fcoords1: First set of fractional coordinates. e.g., [0.5, 0.6,
-            0.7] or [[1.1, 1.2, 4.3], [0.5, 0.6, 0.7]]. It can be a single
-            coord or any array of coords.
-        fcoords2: Second set of fractional coordinates.
-        pbc: a tuple defining the periodic boundary conditions along the three
-            axis of the lattice.
+    Parameters
+    ----------
+    fcoords1 : ArrayLike
+        First set of fractional coordinates. e.g., [0.5, 0.6,
+        0.7] or [[1.1, 1.2, 4.3], [0.5, 0.6, 0.7]]. It can be a single
+        coord or any array of coords.
+    fcoords2 : ArrayLike
+        Second set of fractional coordinates.
+    periodic : ArrayLike
+        a tuple defining the periodic boundary conditions along the three
+        axis of the lattice.
 
-    Returns:
+    Returns
+    -------
+    float
         Fractional distance. Each coordinate must have the  operty that
         abs(a) <= 0.5. Examples:
         pbc_diff([0.1, 0.1, 0.1], [0.3, 0.5, 0.9]) = [-0.2, -0.4, 0.2]
@@ -36,24 +57,23 @@ def pbc_diff_frac_vec(fcoords1: ArrayLike, fcoords2: ArrayLike, periodic):
     fdist = np.subtract(fcoords1, fcoords2)
     return fdist - np.round(fdist) * periodic
 
-
 def pbc_diff_cart(cart_coords1: ArrayLike, cart_coords2: ArrayLike, lattice: Lattice):
     """Returns the Cartesian distance between two coordinates taking into
     account periodic boundary conditions. (from pymatgen)
 
-    Args:
-        fcoords1: First set of fractional coordinates. e.g., [0.5, 0.6,
-            0.7] or [[1.1, 1.2, 4.3], [0.5, 0.6, 0.7]]. It can be a single
-            coord or any array of coords.
-        fcoords2: Second set of fractional coordinates.
-        pbc: a tuple defining the periodic boundary conditions along the three
-            axis of the lattice.
+    Parameters
+    ----------
+    cart_coords1 : ArrayLike
+        First set of Cartesian coordinates
+    cart_coords2 : ArrayLike
+        Second set of Cartesian coordinates
+    lattice : Lattice
+        The Lattice within which the distance should be calculated
 
-    Returns:
[…]

[…]
