+++ src/pylattica/visualization/cell_artist.py
@@ -1,6 +1,6 @@
+from __future__ import annotations
 from abc import abstractmethod
 
-import typing
 from .colors import COLORS
 from ..discrete.state_constants import DISCRETE_OCCUPANCY
 from ..discrete.discrete_step_analyzer import DiscreteStepAnalyzer
@@ -8,10 +8,13 @@ from ..discrete.discrete_state_result_analyzer import DiscreteResultAnalyzer
 from ..core import SimulationResult
 
 from ..core import SimulationState
-from typing import Dict
+from typing import Dict, List, Tuple
 
 
 class CellArtist:
+    """A strategy class for implementing methods of coloring cells in a
+    simulation visualization."""
+
     @abstractmethod
     def get_color_from_cell_state(self, cell_state: Dict):
         pass  # pragma: no cover
@@ -32,39 +35,80 @@ class CellArtist:
 
 
 class DiscreteCellArtist(CellArtist):
+    """A class for coloring cells based on their discrete phase occupancy."""
+
     @classmethod
-    def from_phase_list(cls, phases):
+    def from_phase_list(cls, phases: List[str]) -> DiscreteCellArtist:
+        """Generates a DiscreteCellArtist from a list of phases.
+
+        Parameters
+        ----------
+        phases : List[str]
+            The phases expected in the Simulation.
+
+        Returns
+        -------
+        DiscreteCellArtist
+            The resulting artist.
+        """
         color_map = cls.build_color_map_from_phase_list(phases)
         return cls(color_map)
 
     @classmethod
-    def from_discrete_state(cls, state: SimulationState):
+    def from_discrete_state(cls, state: SimulationState) -> DiscreteCellArtist:
+        """Generates a DiscreteCellArtist from the phases present in the
+        provided simulation state.
+
+        Parameters
+        ----------
+        state : SimulationState
+            The state to generate the artist from.
+
+        Returns
+        -------
+        DiscreteCellArtist
+            The resulting artist.
+        """
         analyzer = DiscreteStepAnalyzer()
         phases = analyzer.phases_present(state)
         return cls.from_phase_list(phases)
 
     @classmethod
-    def from_discrete_result(
-        cls, result: SimulationResult
-    ) -> typing.Dict[str, typing.Tuple[int, int, int]]:
-        """Returns a map of phases to colors that can be used to visualize the phases
+    def from_discrete_result(cls, result: SimulationResult) -> DiscreteCellArtist:
+        """Generates a DiscreteCellArtist from the phases present
+        in a SimulationResult.
 
-        Returns:
-            typing.Dict[str, typing.Tuple[int, int ,int]]: A mapping of phase name to RGB
-            color values
+        Parameters
+        ----------
+        result : SimulationResult
+            The result from which the artist should be generated.
+
+        Returns
+        -------
+        DiscreteCellArtist
+            The resulting artist.
         """
+
         analyzer = DiscreteResultAnalyzer(result)
         return cls.from_phase_list(analyzer.all_phases())
 
     @classmethod
-    def build_color_map_from_phase_list(cls, phases):
-        """Returns a map of phases to colors that can be used to visualize the phases
+    def build_color_map_from_phase_list(
+        cls, phases: List[str]
+    ) -> Dict[str, Tuple[int, int, int]]:
+        """Generates a map of phase name to RGB color from a list of phase names.
 
-        Returns:
-            typing.Dict[str, typing.Tuple[int, int ,int]]: A mapping of phase name to RGB
-            color values
+        Parameters
+        ----------
+        phases : List[str]
+            The phases to include in the color mapping.
+
+        Returns
+        -------
+        Dict[str, Tuple[int, int, int]]
+            A dictionary mapping the phase names to unique color values.
[…]
