+++ src/pylattica/core/simulation_state.py
@@ -26,7 +26,7 @@ class SimulationState:
     @classmethod
     def from_dict(cls, state_dict):
         state = state_dict["state"]
-        state[SITES] = { int(k): v for k, v in state[SITES].items() }
+        state[SITES] = {int(k): v for k, v in state[SITES].items()}
         return cls(state)
 
     @classmethod
@@ -156,7 +156,7 @@ class SimulationState:
         """
 
         if GENERAL in update_batch:
-            for site_id, updates in update_batch.get(SITES, {}).items():
+            for site_id, updates in update_batch[SITES].items():
                 self.set_site_state(site_id, updates)
 
             self.set_general_state(update_batch[GENERAL])
@@ -164,7 +164,7 @@ class SimulationState:
             for site_id, updates in update_batch.items():
                 self.set_site_state(site_id, updates)
 
-    def copy(self) -> "SimulationState":
+    def copy(self) -> SimulationState:
         """Creates a new simulation state identical to this one. This is a deepcopy
         operation, so changing the copy will not change the original.
 
@@ -174,6 +174,6 @@ class SimulationState:
             The copy of this SimulationState
         """
         return SimulationState(self._state)
-    
+
     def __eq__(self, other: SimulationState) -> bool:
         return self._state == other._state
