+++ src/pylattica/structures/honeycomb/lattice.py
@@ -7,6 +7,8 @@ ROOT_3 = math.sqrt(3)
 
 
 class RhombohedralLattice(Lattice):
+    """A helper class for generating rhombohedral lattices."""
+
     def __init__(self):
         lattice_vecs = [
             (HONEYCOMB_SIDE_LENGTH, 0),
