+++ src/pylattica/structures/square_grid/neighborhoods.py
@@ -8,7 +8,16 @@ from ...core.neighborhood_builders import (
 
 
 class VonNeumannNbHood2DBuilder(MotifNeighborhoodBuilder):
+    """A helper class for generating von Neumann type neighborhoods in square 2D structures."""
+
     def __init__(self, size=1):
+        """Constructs the VonNeumannNbHood2DBuilder.
+
+        Parameters
+        ----------
+        size : int, optional
+            The size of the neighborhood, the traditional von Neumann neighborhood would have a size of 1
+        """
         points = get_points_in_cube(-size, size + 1, 2)
 
         filtered_points = []
@@ -20,7 +29,16 @@ class VonNeumannNbHood2DBuilder(MotifNeighborhoodBuilder):
 
 
 class VonNeumannNbHood3DBuilder(MotifNeighborhoodBuilder):
-    def __init__(self, size):
+    """A helper class for generating von Neumann type neighborhoods in square 3D structures."""
+
+    def __init__(self, size: int):
+        """Constructs the VonNeumannNbHood3D Builder
+
+        Parameters
+        ----------
+        size : int
+            The size of the neighborhood.
+        """
         points = get_points_in_cube(-size, size + 1, 3)
 
         filtered_points = []
@@ -32,13 +50,24 @@ class VonNeumannNbHood3DBuilder(MotifNeighborhoodBuilder):
 
 
 class MooreNbHoodBuilder(MotifNeighborhoodBuilder):
+    """A helper class for generating Moore type neighborhoods in square structures."""
+
     def __init__(self, size=1, dim=2):
+        """Constructs the MooreNbHoodBuilder
+
+        Parameters
+        ----------
+        size : int, optional
+            The size of the neighborhood, by default 1
+        dim : int, optional
+            The dimension of the structures to which this builder will apply, by default 2
+        """
         points = get_points_in_cube(-size, size + 1, dim)
         super().__init__(points)
 
 
 class CircularNeighborhoodBuilder(DistanceNeighborhoodBuilder):
-    pass
+    """A semantic class for generating Circular neighborhoods in any structure."""
 
 
 class PseudoHexagonalNeighborhoodBuilder2D(StochasticNeighborhoodBuilder):
@@ -65,6 +94,8 @@ class PseudoHexagonalNeighborhoodBuilder2D(StochasticNeighborhoodBuilder):
 
 
 class PseudoHexagonalNeighborhoodBuilder3D(StochasticNeighborhoodBuilder):
+    """A helper class for generating Pseudohexagonal neighborhoods in 3D."""
+
     def __init__(self):
         common_neighbors = [
             (1, 0, 0),
@@ -100,6 +131,8 @@ class PseudoHexagonalNeighborhoodBuilder3D(StochasticNeighborhoodBuilder):
 
 
 class PseudoPentagonalNeighborhoodBuilder(StochasticNeighborhoodBuilder):
+    """A helper class for generating Pseudopentagonal neighborhoods in 2D."""
+
     def __init__(self):
         motifs = [
             [
