### 2024-02-07 c7c70c2
Remove overkill GoL classes
