+++ tests/discrete/test_discrete_result_analyzer.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pylattica.core import SynchronousRunner
-from pylattica.models.game_of_life import Life
+from pylattica.models.game_of_life import Life, GameOfLifeController
 from pylattica.discrete import PhaseSet, DiscreteResultAnalyzer
 from pylattica.structures.square_grid.grid_setup import DiscreteGridSetup
 
@@ -10,7 +10,8 @@ def discrete_result():
     phases = PhaseSet(["dead", "alive"])
     setup = DiscreteGridSetup(phases)
     simulation = setup.setup_noise(10, ["dead", "alive"])
-    controller = Life(structure = simulation.structure)
+    controller = GameOfLifeController(structure = simulation.structure,
+                                      variant=Life)
     runner = SynchronousRunner(parallel=False)
     return runner.run(simulation.state, controller, 10, verbose=False)
 
@@ -29,7 +30,8 @@ def test_phase_fraction_at_step():
     phases = PhaseSet(["dead", "alive"])
     setup = DiscreteGridSetup(phases)
     simulation = setup.setup_interface(10, "dead", "alive")
-    controller = Life(structure = simulation.structure)
+    controller = GameOfLifeController(structure = simulation.structure,
+                                      variant=Life)
     runner = SynchronousRunner(parallel=False)
     result = runner.run(simulation.state, controller, 10, verbose=False)
 
