### 2024-02-07 cafa2c9
Add note in README about windows
