### 2024-02-08 f29ec84
only remove tmp imgs after gif is saved
