### 2024-02-08 9cde8d2
Add a few more tests
