### 2024-02-08 df911e2
Add detail about neighborhood construction
