### 2024-02-08 65c6790
Add detail about pymatgen lattice functionality
