### 2024-02-08 491bd5e
Remove misleading comment about source of code
