### 2024-02-08 mcgalcode/pylattica#14 @mcgalcode
Hi @amkrajewski,

Thanks for your comments! I have quick responses to a few of them:

pymatgen notices: I added a line in the paper saying that the implementation of methods for converting between lattice coordinates and Cartesian coordinates are taken from pymatgen (I reproduced these in `pylattica`, as opposed to importing them, because in `pymatgen` their implementation is hard-coded to use 3-dimensions, and `pylattica` enforces no such constraint). I think this emphasis reasonably reflects the degree to which `pylattica` relies on `pymatgen`. What do you think? I've also changed the license to match pymatgen's.

L42-46: Can you elaborate on the revision you'd like to see here? I'm trying to communicate that `pylattica` is built for fast prototyping and not for optimizing performance of the resulting simulations. Maybe that doesn't come across?

Neighborhood construction: I agree that there are many additional convenience classes that could be implemented to build neighborhoods! I chose to focus on motifs and radial cutoffs because they are extremely flexible. Every possible neighborhood can be represented by a motif, and radial cutoffs naturally and quickly capture the most common use cases (for instance, basic square grid neighborhood types like von Neumann and Moore are captured by radial cutoffs, and in chemistry, tetrahedral and octahedral coordination environments can be captured this way as well). I've added some of this explanation to the paper. For now, I'd prefer to limit the addition of more neighborhood types until the specific need arises.

Thanks again for your help 🙂
