### 2024-02-08 mcgalcode/pylattica#12 @mcgalcode
Hi again @RiesBen! I've done my best to address your comments here, let me know what you think 🙂
