### 2024-02-08 mcgalcode/pylattica#13 @mcgalcode
Hi @amkrajewski, I've done my best to address your comments here! Let me know what you think (reflected in the master branch, and the paper branch).
