### 2024-02-08 mcgalcode/pylattica#14 @amkrajewski
Hi @mcgalcode !

### pymatgen:
In the revision I see it says:
> These structures are implemented using a `Lattice` class which uses functionality from `pymatgen` to convert coordinates between the basis of the lattice and the Cartesian basis.

which doesn't really carry the message. At the same time, your comment here explains it perfectly! With slight modification:

> These structures are implemented using a `Lattice` class which was cloned from `pymatgen` and then adjusted to the needs of `pylattica`, while mostly retaining compatibility,  primarily because current `pymatgen`'s implementation is hard-coded to use 3-dimensions, while `pylattica` strives for generality and enforces no such constraint.

Hope that helps :)

### L42-46:
I like the content. I just meant that a style revision would be good here. In first sentence you say that performance was not the priority and in the second one that its good for applications where performance is not important. It felt a bit redundant to me.

### Neighborhood construction:
Looks good to me!
