### 2024-02-09 mcgalcode/pylattica#13 @amkrajewski
Hi @mcgalcode! The changes you've made look great! I think it's pretty much ready! Let me have a bit closer look tomorrow, but all of my original concerns are addressed.

I caught a couple minor things looking through, but fixing them shouldn't take more than a few minutes:
- For the development install instructions in README, you may want to add `-e`, so that it installs in editable mode and changes when users play with it.
- You may want to consider removing CHANGELOG file (empty now) and writing release notes through GitHub. It's way more pleasant to maintain and notifies people when you release. Also, the changelog on PyPI (set in `pyproject.toml`) is pointing to not existing doc page.
- The `setuptools` version required by build system should likely be 61 or higher based on project structure.
- The date in license is old I think.
- The license on PyPI (set in `pyproject.toml`) is not MIT.
