### 2024-02-09 1af83d1
Small tweaks to license and installation
