### 2024-02-09 mcgalcode/pylattica#13 @mcgalcode
Hi @amkrajewski ! Ah thanks, these are some good catches. I fixed them all (and agree re: removing CHANGELOG), except for the first - I've got `pip install -e . ` under the *Development* section.
