### 2024-02-09 ed16a38
Rephrase paper sentence about pymatgen lattices
