### 2024-02-09 mcgalcode/pylattica#14 @mcgalcode
Hi @amkrajewski,

re: pymatgen, Thanks, that makes a lot of sense! I took your suggestion and tweaked it slightly because the `pylattica.core.Lattice` is not meant to be interoperable with the `pymatgen.Lattice` class (the connection is via their respective `Structure` classes), so I don't want to give that impression (though this is a relatively subtle distinction).

I tried to reword 42-46 as well - does that sound better?
