### 2024-02-09 mcgalcode/pylattica#14 @amkrajewski
Both look great!

I think the flowchart is the only missing piece. In case you didn't start working on it yet, you can try: https://github.com/jgraph/drawio-desktop
It is FOSS and has a neat functionality of embedding chart source into PNG, so you or anyone else can easily adopt or modify it later.
