### 2024-02-09 e017ec5
Add flowchart section to paper
