### 2024-02-09 mcgalcode/pylattica#14 @mcgalcode
That was a great suggestion! I just added a flowchart to the manuscript now.
