### 2024-02-09 mcgalcode/pylattica#13 @amkrajewski
All looks good!
