### 2024-02-10 mcgalcode/pylattica#14 @amkrajewski
You may look into adding a bit more information into the flowchart on _how_ the entities are related and _why_.

At the same time, the current version does its job, so you can consider my concerns resolved 👍
