### 2024-02-12 mcgalcode/pylattica#13 @mcgalcode
Thanks for all these helpful comments @amkrajewski!
