### 2024-02-12 mcgalcode/pylattica#14 @mcgalcode
@amkrajewski Thanks for your help here - I will brainstorm ways to improve that flowchart for the documentation website :)
