### 2024-02-13 mcgalcode/pylattica#12 @mcgalcode
Hey @RiesBen just want to bump here in case you didn't see this one!
