### 2024-03-05 mcgalcode/pylattica#12 @RiesBen
awesome looks good to me :)
