+++ pyproject.toml
@@ -36,7 +36,6 @@ requires-python = '>=3.8'
 dependencies = [
     'numpy >= 1.21.5',
     'matplotlib >= 3.5.1',
-    'scipy >= 1.8.0',
     'tqdm >= 4.63.0',
     'pymatgen >= 2022.3.29',
     'plotly >= 5.6.0',
