### 2024-03-13 #6170 @mcgalcode
Hi @richardjgowers, I've completed the tasks listed above.

The current version is v0.1.3 (https://github.com/mcgalcode/pylattica/releases/tag/v0.1.3, https://pypi.org/project/pylattica/)

The zenodo DOI is here: 10.5281/zenodo.10815119
