### 2024-03-18 d3912d0
Pass kwargs to get images
