### 2024-03-18 69db8ae
Better from dict
