### 2024-05-06 #6170 @mcgalcode
Hi @Kevin-Mattheus-Moerman, thanks for your suggestions! I've updated the affiliations as you suggested, and added departments as well.

That DOI isn't correct though - it seems to point to a textbook, and my citation is for an issue of Scientific American.
